requirements.txt
gspread
oauth2client
python-dotenv
python-telegram-bot
Unidecode
psycopg2
psycopg2-binary
tk
playsound